# Reference count TokenStream in fallback mode